Rewrite pull method and stuff using pulled data
This is totally rewritting the pull method and everything that uses data
pulled from spreadsheet (so practically, the whole tool).

Nowadays, it will pull 30
rows, process and pull next rows. This is optimizing API calls.

Also, this protects us from need to have millions of columns describing
possible questions. We only need to have a question to keyword map.

Bug: T207779
Change-Id: I705c60ab4e6d653f5a3805d017632fcf8234a1a9